# core: reimplement Metadata to avoid allocation.