Key things:

* This moves metadata closer to being in the same format as Netty's headers.
* Expectation is that there will be very few or no headers, so use linear time algorithms
* All validation is done in the marshallers.
* Keep a reference to the object.  This will be used when the eventual canonicalization happens.

Proper benchmarks to come, initial results look good.