@adriancole 2016-09-15
unsurprised about the good scores. Since there are folks maniacal about performance in the house, the more explicit code isn't likely at risk of people not understanding it. looks good!