@buchgr 2016-09-15
@carl-mastrangelo memory footprint / savings would also be interesting :).