@adriancole 2016-09-16
so what are you thinking? make a char array of equal length and copy into
that assuming ascii, and if it isn't printable substitute 'X' or something?

On Fri, Sep 16, 2016 at 11:13 AM, Eric Anderson <notifications@github.com>
wrote:

> *@ejona86* commented on this pull request.
> ------------------------------
>
> In core/src/main/java/io/grpc/Metadata.java
> <https://github.com/grpc/grpc-java/pull/2274>:
>
> >    public String toString() {
> -    return "Metadata(" + toStringInternal() + ")";
> -  }
> -
> -  private String toStringInternal() {
> -    return store.toString();
> +    StringBuilder sb = new StringBuilder("Metadata(");
> +    for (int i = 0; i < size; i++) {
> +      if (i != 0) {
> +        sb.append(',');
> +      }
> +      sb.append(new String(name(i), 0)).append('=');
> +      if (typedValues[i] != null) {
> +        sb.append(typedValues[i]);
> +      } else {
> +        sb.append(new String(value(i), 0));
>
> Not outputting control characters is very, very valuable. If displayed on
> a console, it would trash the console. It doesn't have to be hex, but raw
> binary in a string is bad.
>
> I 100% expect we will want to be able to read those binary values from log
> output, especially after errors.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/grpc/grpc-java/pull/2274>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAD617zXMD-I7k2KBiKHmpMO0As-xP1hks5qqglsgaJpZM4J9Zbf>
> .
>